requirements.txt
Django==1.10.7
django-filter==1.0.2
djangorestframework==3.6.2
Pillow==4.1.1
celery==4.0.2
pika==0.10.0